requirements.txt
pyperclip
waapi-client